Feature/rahadov

+++ TransactionFlow.BillingSystem/Controllers/AccountController.cs
@@ -12,21 +12,32 @@ public class AccountController:ControllerBase
 {
     private ITransferService _transferService;
     private IAccountService _accountService;
+    private ISessionService _sessionService;
 
-    public AccountController(ITransferService transferService, IAccountService accountService)
+    public AccountController(ITransferService transferService, IAccountService accountService, ISessionService sessionService)
     {
         _transferService = transferService;
         _accountService = accountService;
+        _sessionService = sessionService;
     }
     
     [Route(nameof(CreateAccount))]
     [HttpPost]
-    public IActionResult CreateAccount(int customerId)
+    public IActionResult CreateAccount([FromBody] int customerId)
     {
-        var result = _accountService.CreateAccount(customerId);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = _accountService.CreateAccount(customerId);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
@@ -34,12 +45,21 @@ public class AccountController:ControllerBase
 
     [Route(nameof(DeleteAccountAsync))]
     [HttpPost]
-    public async Task<IActionResult> DeleteAccountAsync(int accountId)
+    public async Task<IActionResult> DeleteAccountAsync([FromBody] int accountId)
     {
-        var result = await _accountService.DeleteAccountAsync(accountId);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = await _accountService.DeleteAccountAsync(accountId);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
@@ -47,12 +67,21 @@ public class AccountController:ControllerBase
     
     [Route(nameof(DeactivateAccountAsync))]
     [HttpPost]
-    public async Task<IActionResult> DeactivateAccountAsync(int accountId)
+    public async Task<IActionResult> DeactivateAccountAsync([FromBody] int accountId)
     {
-        var result = await _accountService.DeactivateAccountAsync(accountId);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = await _accountService.DeactivateAccountAsync(accountId);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
@@ -60,12 +89,21 @@ public class AccountController:ControllerBase
     
     [Route(nameof(ActivateAccountAsync))]
     [HttpPost]
-    public async Task<IActionResult> ActivateAccountAsync(int accountId)
+    public async Task<IActionResult> ActivateAccountAsync([FromBody] int accountId)
     {
-        var result = await _accountService.ActivateAccountAsync(accountId);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = await _accountService.ActivateAccountAsync(accountId);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
@@ -75,12 +113,21 @@ public class AccountController:ControllerBase
     //Deducts amount + fee from sender customer's main account and adds to receiver customer's main account.
     [Route(nameof(TransferMoneyCToCAsync))]
     [HttpPost]
-    public async Task<IActionResult> TransferMoneyCToCAsync(TransferDto transferDto)
+    public async Task<IActionResult> TransferMoneyCToCAsync([FromBody] TransferDto transferDto)
     {
-        var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.CToC);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.CToC);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
@@ -91,12 +138,21 @@ public class AccountController:ControllerBase
     //receiver customer's certain account.
     [Route(nameof(TransferMoneyAToAAsync))]
     [HttpPost]
-    public async Task<IActionResult> TransferMoneyAToAAsync(TransferDto transferDto)
+    public async Task<IActionResult> TransferMoneyAToAAsync([FromBody] TransferDto transferDto)
     {
-        var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.AToA);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.AToA);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
@@ -107,12 +163,21 @@ public class AccountController:ControllerBase
     //receiver customer's certain account.
     [Route(nameof(TransferMoneyCToAsync))]
     [HttpPost]
-    public async Task<IActionResult> TransferMoneyCToAsync(TransferDto transferDto)
+    public async Task<IActionResult> TransferMoneyCToAsync([FromBody] TransferDto transferDto)
     {
-        var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.CToA);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.CToA);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
@@ -123,12 +188,21 @@ public class AccountController:ControllerBase
     //receiver customer's main account.
     [Route(nameof(TransferMoneyAToCAsync))]
     [HttpPost]
-    public async Task<IActionResult> TransferMoneyAToCAsync(TransferDto transferDto)
+    public async Task<IActionResult> TransferMoneyAToCAsync([FromBody] TransferDto transferDto)
     {
-        var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.AToC);
-        if (result.IsFailed)
+        var key = Request.Headers["Idempotency-key"].ToString();
+
+        var idempotencyResult = _sessionService.Get(key);
+        if (idempotencyResult == null)
         {
-            return BadRequest(result.Reasons);
+            var result = await _transferService.TransferMoneyAsync(transferDto,TransferConditions.AToC);
+            if (result.IsFailed)
+            {
+                return BadRequest(result.Reasons);
+            }
+
+            _sessionService.Set(key);
+            return Ok();
         }
 
         return Ok();
